pyproject.toml
[project]
name = "clinical-data-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fhir-resources>=8.2.0",
    "sqlalchemy>=2.0.49",
]

[tool.ruff]
line-length = 120
